Single arity template handlers are deprecated
The template handler now allows for the new Rails 6 api taking in both
the template object and a source argument, but still allows the old
behavior.

Fixes:

DEPRECATION WARNING: Single arity template handlers are deprecated. Template handlers must
now accept two parameters, the view object and the source for the view object.
Change:
  >> Opal::Rails::TemplateHandler.call(template)
To:
  >> Opal::Rails::TemplateHandler.call(template, source)
 (called from <main> at /Users/elia/Code/foglio_appuntamenti/config/application.rb:20)